+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import Practice4 from './templates/practice4';
+import Practice5 from './templates/practice5';
 
 function App() {
   return (
@@ -7,7 +7,7 @@ function App() {
       <header className="bg-[#94A3B8] text-center p-[20px] text-4xl text-[#F9FAFB]">React-v3</header>
       <main className="mt-[100px]">
         <section className="max-w-[600px] mx-auto">
-          <Practice4 />
+          <Practice5 />
         </section>
       </main>
     </>
